factor out chat mod
